requirements.txt
streamlit>=1.30.0
anthropic>=0.25.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
readability-lxml>=0.8.1
weasyprint>=61.0
jinja2>=3.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
pytest>=8.0.0
groq>=0.9.0
